backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
httpx==0.27.2
redis==5.1.0
stripe==10.12.0
anthropic==0.34.2
resend==2.4.0
python-dateutil==2.9.0
psycopg2-binary==2.9.9
# Iteración 2 — IA clínica
faster-whisper==1.0.3
aiofiles==23.2.1
